Add edge/vertex label mapping functions
I wanted to put these in the appropriate type classes
with the vertex/edge inspectors, but I don't think the
types can work out...

Something like:

> class HasVertexLabel g where
>   type VertexLabel g
>   mapVertexLabel :: (HasVertexLabel g') => g -> (VertexLabel g -> VertexLabel g') -> g'

does not work because it requires a level of polymorphism not possible
(since it requires that mapVertexLabel be able to produce a different
underlying graph type from g).  This is where the representation in FGL
is a bit more convenient since you have a single type parameter you
can change easily.

This commit just puts the functions (not in type classes)
in the relevant adapter modules, which have the same type of
polymorphism as FGL.  That should be fine, though it is a bit less
convenient than having the functions in type classes.