Clean up the top level module
The top level module will be the new entry point that can be imported to
expose almost all library functionality.  It still needs the mutable
graph constructors and the various graph data types.  This requires that
the classes be split into a separate module so that they can be imported
by the graph implementations without introducing cycles.