Add Functor instances for PatriciaTree and Ctx
`Bifunctor` now has a quantified `Functor` superclass in recent versions of
`base`, so we need to define a `Functor` instance for `PatriciaTree` to go
along with its existing `Bifunctor` instance. The easiest way to do so is via
`DeriveFunctor`. This entails deriving a `Functor` instance for `Ctx` (on top
of which `PatriciaTree` is defined) as well.